Update takeover.rs
`match` statements with only one useful arm can be reduced to an `if-else` statement. This helps improve readability, an `if-else` statement nests lesser than a `match`.

Use the pattern of the single match arm as the pattern within an `if let` statement. The body of the wildcard arm resides within the `else` block.